refactor: memory
