+++ cookbook/agent_concepts/memory/v2/long_term_memory.py
@@ -1,6 +1,6 @@
 from agno.agent.agent import Agent
 from agno.db.postgres import PostgresDb
-from agno.memory.memory import Memory
+from agno.memory.manager import MemoryManager  # noqa: F401
 from agno.models.openai import OpenAIChat
 
 db_url = "postgresql+psycopg://ai:ai@localhost:5532/ai"
@@ -11,13 +11,25 @@ db = PostgresDb(
     user_memory_table="user_memories",
 )
 
-memory = Memory(db=db)
-
+# 1. Using the default memory manager by setting enable_user_memories=True
 agent = Agent(
     model=OpenAIChat(id="gpt-4o-mini"),
-    memory=memory,
+    db=db,
     session_id="long_term_memory",
     enable_user_memories=True,
 )
 
 agent.print_response("I love astronomy, specifically the science behind nebulae")
+
+# 2. Using a custom memory manager by creating your own instance of MemoryManager
+# memory_manager = MemoryManager(model=OpenAIChat(id="gpt-4o-mini"), db=db)
+
+# agent = Agent(
+#     model=OpenAIChat(id="gpt-4o-mini"),
+#     db=db,
+#     session_id="long_term_memory",
+#     enable_user_memories=True,
+#     memory_manager=memory_manager,
+# )
+
+# agent.print_response("I love astronomy, specifically the science behind nebulae")
