+++ libs/agno/agno/db/base.py
@@ -129,6 +129,11 @@ class BaseDb(ABC):
     ) -> Optional[Union[UserMemory, Dict[str, Any]]]:
         raise NotImplementedError
 
+    # TODO:
+    # @abstractmethod
+    # def clear_memories(self) -> None:
+    #     raise NotImplementedError
+
     # --- Metrics ---
     @abstractmethod
     def get_metrics(
